Platform: Move Collectibles to separate package
