+++ Platform/src/platform/collectible/CollectResult.java
@@ -1,4 +1,4 @@
-package platform.component;
+package platform.collectible;
 
 /**
  *
